Moving mAuth to LoginActivity
